+++ composer.json
@@ -2,7 +2,7 @@
   "name": "optimistex/yii2-expression",
   "description": "Taken from http://www.phpclasses.org/browse/file/11680.html, cred to Miles Kaufmann",
   "type": "yii2-extension",
-  "version": "1.1.2",
+  "version": "2.0",
   "license": "MIT",
   "authors": [
     {
