+++ tests/Expression.php
@@ -3,11 +3,12 @@
 use PHPUnit\Framework\TestCase;
 
 
-
-class ExpressionTest extends TestCase {
-    public function arrayTest($array) {
+class ExpressionTest extends TestCase
+{
+    public function arrayTest($array)
+    {
         $expr = new Expression();
-        for ($i=0; $i<count($array); $i++) {
+        for ($i = 0; $i < count($array); $i++) {
             $result = $expr->evaluate($array[$i]);
             $this->assertEquals($result, eval("return " . $array[$i] . ";"));
         }
@@ -20,108 +21,125 @@ class ExpressionTest extends TestCase {
     }
     */
     // -------------------------------------------------------------------------
-    public function testIntegers() {
+    public function testIntegers()
+    {
         $ints = array("100", "3124123", (string)PHP_INT_MAX, "-1000");
         $expr = new Expression();
-        for ($i=0; $i<count($ints); $i++) {
+        for ($i = 0; $i < count($ints); $i++) {
             $result = $expr->evaluate($ints[$i]);
             $this->assertEquals($result, intval($ints[$i]));
         }
 
     }
+
     // -------------------------------------------------------------------------
-    public function testFloats() {
+    public function testFloats()
+    {
         $ints = array("10.10", "0.01", ".1", "1.", "-100.100", "1.10e2", "-0.10e10");
         $expr = new Expression();
-        for ($i=0; $i<count($ints); $i++) {
+        for ($i = 0; $i < count($ints); $i++) {
             $result = $expr->evaluate($ints[$i]);
             $this->assertEquals($result, floatval($ints[$i]));
         }
     }
+
     // -------------------------------------------------------------------------
-    public function testAritmeticOperators() {
+    public function testAritmeticOperators()
+    {
         $expressions = array("20+20", "-20+20", "-0.1+0.1", ".1+.1", "1.+1.",
-                             "0.1+(-0.1)", "20*20", "-20*20", "20*(-20)", "1.*1.",
-                             ".1*.1", "20-20", "-20-20", "20/20", "-20/20", "10%20",
-                             "10%9", "20%9");
+            "0.1+(-0.1)", "20*20", "-20*20", "20*(-20)", "1.*1.",
+            ".1*.1", "20-20", "-20-20", "20/20", "-20/20", "10%20",
+            "10%9", "20%9");
         $this->arrayTest($expressions);
         try {
             $expr = new Expression();
             $expr->evaluate('10/0');
             $this->assertTrue(false); // will fail if evaluate don't throw exception
-        } catch(Exception $e) {
+        } catch (Exception $e) {
             $this->assertTrue(true);
         }
     }
+
     // -------------------------------------------------------------------------
-    public function testSemicolon() {
+    public function testSemicolon()
+    {
         $expr = new Expression();
         $result = $expr->evaluate("10+10;");
         $this->assertEquals($result, "20");
     }
+
     // -------------------------------------------------------------------------
-    public function testBooleanComparators() {
+    public function testBooleanComparators()
+    {
         $expressions = array("10 == 10", "10 == 20", "0.1 == 0.1", "0.1 == 0.2",
-                             "10 != 10", "20 != 10", "0.1 != 0.1", "0.1 != 0.2",
-                             "10 < 10", "20 < 10", "10 < 20", "0.1 < 0.2",
-                             "0.2 < 0.1", "0.1 < 0.1", "10 > 10", "20 > 10",
-                             "10 > 20", "0.1 > 0.2", "0.2 > 0.1", "0.1 > 0.1",
-                             "10 <= 10", "20 <= 10", "10 <= 20", "0.1 <= 0.2",
-                             "0.2 <= 0.1", "0.1 <= 0.1", "10 >= 10", "20 >= 10",
-                             "10 >= 20", "0.1 >= 0.2", "0.2 >= 0.1", "0.1 >= 0.1");
+            "10 != 10", "20 != 10", "0.1 != 0.1", "0.1 != 0.2",
+            "10 < 10", "20 < 10", "10 < 20", "0.1 < 0.2",
+            "0.2 < 0.1", "0.1 < 0.1", "10 > 10", "20 > 10",
+            "10 > 20", "0.1 > 0.2", "0.2 > 0.1", "0.1 > 0.1",
+            "10 <= 10", "20 <= 10", "10 <= 20", "0.1 <= 0.2",
+            "0.2 <= 0.1", "0.1 <= 0.1", "10 >= 10", "20 >= 10",
+            "10 >= 20", "0.1 >= 0.2", "0.2 >= 0.1", "0.1 >= 0.1");
         $this->arrayTest($expressions);
     }
+
     // -------------------------------------------------------------------------
-    public function testBooleanOperators() {
+    public function testBooleanOperators()
+    {
         $expressions = array("10 == 10 && 10 == 10", "10 != 10 && 10 != 10",
-                             "10 == 20 && 10 == 10", "10 == 10 && 10 == 20",
-                             "0.1 == 0.1 && 0.1 == 0.1", "0.1 == 0.2 && 0.1 == 0.1",
-                             "0.1 == 0.1 && 0.1 == 0.2", "10 == 10 || 10 == 10",
-                             "10 == 20 || 10 == 10", "10 == 10 || 10 == 20",
-                             "0.1 == 0.1 || 0.1 == 0.1", "0.1 == 0.2 || 0.1 == 0.1",
-                             "0.1 == 0.1 || 0.1 == 0.2");
+            "10 == 20 && 10 == 10", "10 == 10 && 10 == 20",
+            "0.1 == 0.1 && 0.1 == 0.1", "0.1 == 0.2 && 0.1 == 0.1",
+            "0.1 == 0.1 && 0.1 == 0.2", "10 == 10 || 10 == 10",
+            "10 == 20 || 10 == 10", "10 == 10 || 10 == 20",
+            "0.1 == 0.1 || 0.1 == 0.1", "0.1 == 0.2 || 0.1 == 0.1",
+            "0.1 == 0.1 || 0.1 == 0.2");
         $this->arrayTest($expressions);
         $expressions = array('("foo" == "foo") && "a" || "b"' => "a",
-                             '("foo" == "bar") && "a" || "b"' => "b");
+            '("foo" == "bar") && "a" || "b"' => "b");
         $expr = new Expression();
         foreach ($expressions as $expression => $value) {
             $result = $expr->evaluate($expression);
             $this->assertEquals($result, $value);
         }
     }
+
     // -------------------------------------------------------------------------
-    public function testKeywords() {
+    public function testKeywords()
+    {
         $expressions = array("1 == true", "true == true", "false == false",
-                             "false != true", "null == null", "null != true");
+            "false != true", "null == null", "null != true");
         $expr = new Expression();
         foreach ($expressions as $expression) {
             $result = $expr->evaluate($expression);
             $this->assertTrue((bool)$result);
         }
-        $expressions = array("foo = true"=>true, "foo = false"=>false, "foo = null"=>null);
+        $expressions = array("foo = true" => true, "foo = false" => false, "foo = null" => null);
         foreach ($expressions as $expression => $value) {
             $expr = new Expression();
-            $result = $expr->evaluate($expression);
+            $expr->evaluate($expression);
             $result = $expr->evaluate("foo");
             $this->assertEquals($result, $value);
         }
-        
+
     }
+
     // -------------------------------------------------------------------------
-    public function testNegation() {
+    public function testNegation()
+    {
         $expressions = array("!(10 == 10)", "!1", "!0");
         $this->arrayTest($expressions);
     }
+
     // -------------------------------------------------------------------------
-    public function testStrings() {
+    public function testStrings()
+    {
         $expressions = array('"foo" == "foo"', '"foo\\"bar" == "foo\\"bar"',
-                             '"f\\"oo" != "f\\"oo"', '"foo\\"" != "foo\\"bar"',
-                             "'foo\"bar' == 'foo\"bar'", "'foo' == 'foo'",
-                             "'foo\\'foo' != 'foo'", '"foo\\\\" == "foo\\\\"',
-                             "'foo\\\\' == 'foo\\\\'");
+            '"f\\"oo" != "f\\"oo"', '"foo\\"" != "foo\\"bar"',
+            "'foo\"bar' == 'foo\"bar'", "'foo' == 'foo'",
+            "'foo\\'foo' != 'foo'", '"foo\\\\" == "foo\\\\"',
+            "'foo\\\\' == 'foo\\\\'");
         $this->arrayTest($expressions);
         $expressions = array('"foo" + "bar"' => 'foobar', "'foo' + 'bar'" => 'foobar',
-                             '"foo\\"bar" + "baz"' => "foo\"barbaz");
+            '"foo\\"bar" + "baz"' => "foo\"barbaz");
         $expr = new Expression();
         foreach ($expressions as $expression => $value) {
             $result = $expr->evaluate($expression);
@@ -130,13 +148,15 @@ class ExpressionTest extends TestCase {
         $result = $expr->evaluate('"foo\"ba\\\\\\"r" =~ "/foo/"');
         $this->assertEquals((boolean)$result, true);
     }
+
     // -------------------------------------------------------------------------
-    public function testMatchers() {
+    public function testMatchers()
+    {
         $expressions = array('"foobar" =~ "/([fo]+)/"' => 'foo',
-                             '"foobar" =~ "/([0-9]+)/"' => null,
-                             '"1020" =~ "/([0-9]+)/"'=> '1020',
-                             '"1020" =~ "/([a-z]+)/"' => null);
-        
+            '"foobar" =~ "/([0-9]+)/"' => null,
+            '"1020" =~ "/([0-9]+)/"' => '1020',
+            '"1020" =~ "/([a-z]+)/"' => null);
+
         foreach ($expressions as $expression => $group) {
             $expr = new Expression();
             $result = $expr->evaluate($expression);
@@ -148,30 +168,34 @@ class ExpressionTest extends TestCase {
             }
         }
     }
+
     // -------------------------------------------------------------------------
-    public function testVariableAssignment() {
+    public function testVariableAssignment()
+    {
         $expressions = array('foo = "bar"' => array('var' => 'foo', 'value' => 'bar'),
-                             'foo = 10' => array('var' => 'foo', 'value' => 10),
-                             'foo = 0.1' => array('var' => 'foo', 'value' => 0.1),
-                             'foo = 10 == 10' => array('var' => 'foo', 'value' => 1),
-                             'foo = 10 != 10' => array('var' => 'foo', 'value' => 0),
-                             'foo = "foo" =~ "/[fo]+/"' => array('var' => 'foo', 'value' => 1),
-                             'foo = 10 + 10' => array('var' => 'foo', 'value' => 20));
+            'foo = 10' => array('var' => 'foo', 'value' => 10),
+            'foo = 0.1' => array('var' => 'foo', 'value' => 0.1),
+            'foo = 10 == 10' => array('var' => 'foo', 'value' => 1),
+            'foo = 10 != 10' => array('var' => 'foo', 'value' => 0),
+            'foo = "foo" =~ "/[fo]+/"' => array('var' => 'foo', 'value' => 1),
+            'foo = 10 + 10' => array('var' => 'foo', 'value' => 20));
         foreach ($expressions as $expression => $object) {
             $expr = new Expression();
             $expr->evaluate($expression);
             $this->assertEquals($expr->evaluate($object['var']), $object['value']);
         }
     }
+
     // -------------------------------------------------------------------------
-    public function testJSON() {
+    public function testJSON()
+    {
         $expressions = array(
             array("foo" => "bar"),
             array('foo\\"bar' => "baz"),
             array("foo}" => "bar"),
-            array(10,20,30,40),
-            array(10,"]",30),
-            array(10, array("foo"=>"bar"), 30)
+            array(10, 20, 30, 40),
+            array(10, "]", 30),
+            array(10, array("foo" => "bar"), 30)
         );
         $expr = new Expression();
         foreach ($expressions as $expression) {
@@ -201,39 +225,42 @@ class ExpressionTest extends TestCase {
         $result = $expr->evaluate('foo["foo"]');
         $this->assertEquals($result, 'bar');
     }
+
     // -------------------------------------------------------------------------
-    public function testCustomFunctions() {
+    public function testCustomFunctions()
+    {
         $functions = array('square(x) = x*x' => array(
-                                'square(10)' => 100,
-                                'square(10) == 100' => 1
-                           ),
-                           'string() = "foo"' => array(
-                                'string() =~ "/[fo]+/"' => 1,
-                                'string() == "foo"' => 1,
-                                'string() != "bar"' => 1
-                           ),
-                           'number(x) = x =~ "/^[0-9]+$/"' => array(
-                                'number("10")' => 1,
-                                'number("10foo")' => 0
-                           ),
-                           'logic(x, y) = x == "foo" || x == "bar"' => array(
-                                'logic( "foo", 1 )' => 1,
-                                'logic("bar", 1)' => 1,
-                                'logic("lorem", 1)' => 0
-                           ));
+            'square(10)' => 100,
+            'square(10) == 100' => 1
+        ),
+            'string() = "foo"' => array(
+                'string() =~ "/[fo]+/"' => 1,
+                'string() == "foo"' => 1,
+                'string() != "bar"' => 1
+            ),
+            'number(x) = x =~ "/^[0-9]+$/"' => array(
+                'number("10")' => 1,
+                'number("10foo")' => 0
+            ),
+            'logic(x, y) = x == "foo" || x == "bar"' => array(
+                'logic( "foo", 1 )' => 1,
+                'logic("bar", 1)' => 1,
+                'logic("lorem", 1)' => 0
+            ));
         foreach ($functions as $function => $object) {
             $expr = new Expression();
             $expr->evaluate($function);
             foreach ($object as $fn => $value) {
                 $this->assertEquals($expr->evaluate($fn), $value);
             }
-        }       
+        }
     }
 
     // -------------------------------------------------------------------------
-    public function testCustomClosures() {
+    public function testCustomClosures()
+    {
         $expr = new Expression();
-        $expr->functions['even'] = function($a) {
+        $expr->functions['even'] = function ($a) {
             return $a % 2 == 0;
         };
         $values = array(10 => true, 20 => true, 1 => false, 3 => false, 4 => true);
@@ -243,5 +270,3 @@ class ExpressionTest extends TestCase {
     }
     // */
 }
-
-?>
